+++ src/main/kotlin/calculator/OperatorType.kt
@@ -1,0 +1,5 @@
+package calculator
+
+enum class OperatorType(val symbol: String) {
+    PLUS("+"), SUBTRACT("-"), MULTIPLY("*"), DIVIDE("/")
+}
