+++ src/test/kotlin/calculator/CalculatorTest.kt
@@ -1,3 +1,5 @@
+package calculator
+
 import io.kotest.assertions.throwables.shouldNotThrowAny
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
