+++ src/test/kotlin/person/PersonFunTest.kt
@@ -1,3 +1,5 @@
+package person
+
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
