+++ src/test/kotlin/person/PersonKoTest.kt
@@ -1,3 +1,5 @@
+package person
+
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.shouldBe
